#46/스폰서 페이지를 이벤트 정보 페이지로 이동
