+++ androidapp/app/src/main/java/com/droidknights/app2020/di/ActivityBindingModule.kt
@@ -6,7 +6,6 @@ import com.droidknights.app2020.ui.info.InfoModule
 import com.droidknights.app2020.ui.MainActivityModule
 import com.droidknights.app2020.ui.mypage.MypageModule
 import com.droidknights.app2020.ui.schedule.ScheduleModule
-import com.droidknights.app2020.ui.sponsor.SponsorModule
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
@@ -21,8 +20,7 @@ abstract class ActivityBindingModule {
             MainActivityModule::class,
             ScheduleModule::class,
             InfoModule::class,
-            MypageModule::class,
-            SponsorModule::class
+            MypageModule::class
         ]
     )
     internal abstract fun getMainActivity() : MainActivity
